[virtio harness] enable fuzzer for virtio network boot (PXEv4)
With this commit we have a working fuzzer example for the PXEv4 network
boot (i.e. boot via Virtio network device).
The fuzzer payload is injected into a filebuffer that normally would
hold a bootable PE image to execute. Hence, the buffer input will be
interpreted as such as well and it is attempted to load/boot this
injected code.
It does not provide actual useful results because so far in all cases
loading got prevented by TCG Trusted Boot.