[agent] decouple agent from harnesses
Restores agent implementation that is independent from implementation of
fuzzing harnesses.

Harness configuration flags were removed, and use of memory allocation
that was dependent on TdHob harness now depends on general
EARLY_BOOT_FUZZING flag.
This flag is not set by default, and can be set by specific harnesses
(such as TdHob harness) that need to be implemented in early boot code
(such as TdxStartup.c).

The agent implementation in this branch serves as a clean basis state,
from which individual harnesses can be developed in the future.